handled authorization in creating categories

+++ Controllers/categoryController.js
@@ -6,10 +6,17 @@ const Categories = require("../Models/Categories");
 
 const getAllCategories = async (req, res, next) => {
   try {
-    const categories = await Categories.find();
-    if (categories.length == 0)
-      return next(new AppError("no categories found!"));
-    res.send({ message: "All categories retrieved successfully", categories });
+    if (req.authorizedUser.role == "admin") {
+      const categories = await Categories.find();
+      if (categories.length == 0)
+        return next(new AppError("no categories found!"));
+      res.send({
+        message: "All categories retrieved successfully",
+        categories,
+      });
+    } else {
+      res.send("you are not an admin");
+    }
   } catch (error) {
     return next(error);
   }
@@ -19,11 +26,15 @@ const getAllCategories = async (req, res, next) => {
 
 const createCategory = async (req, res, next) => {
   try {
-    const { name } = req.body;
-    if (!name) return next(new AppError("Please enter the category name!"));
-    const category = new Categories({ name });
-    await category.save();
-    res.send({ message: "Category created successfully", category });
+    if (req.authorizedUser.role == "admin") {
+      const { name } = req.body;
+      if (!name) return next(new AppError("Please enter the category name!"));
+      const category = new Categories({ name });
+      await category.save();
+      res.send({ message: "Category created successfully", category });
+    } else {
+      res.send("you are not an admin");
+    }
   } catch (error) {
     return next(error);
   }
@@ -33,47 +44,45 @@ const createCategory = async (req, res, next) => {
 
 const updateCategoryById = async (req, res, next) => {
   try {
-    const category = await Categories.findById(req.params.id);
-    if (!category) return next(new AppError("this category does not exist"));
-    const newCategory = await Categories.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    // if (req.authorizedUser.id == category.userId) 
-    // {
-    // } 
-    // else
-    //  {
-    //   res.send({ message: "you can't edit other users posts" });
-    // }
-    res.send({ message: "Category updated successfully", newCategory });
-  } catch (err) 
-  {
-    return next(err);
-  }
-};
-
-
-//http://localhost:8080/categories/:id
-
-const deleteCategoryById = async (req, res, next) => {
-  try {
-    const categories = await Categories.findById(req.params.id);
-    if (!categories) return next(new AppError("category does not exist"));
-    // if (
-    //   req.authorizedUser.id == categories.userId ||
-    //   req.authorizedUser.role == "admin"
-    // ) 
-    // {
-      await Categories.findByIdAndDelete(req.params.id);
-      res.send({ message: "Category deleted successfully" });
-    // } 
-    // else 
-    // {
-    //   res.send({ message: "you can't delete other users posts" });
-    // }
+    if (req.authorizedUser.role == "admin") {
+      const category = await Categories.findById(req.params.id);
+      if (!category) return next(new AppError("this category does not exist"));
+      const newCategory = await Categories.findByIdAndUpdate(
+        req.params.id,
+        req.body,
+        {
+          new: true,
+        }
+      );
+      res.send({ message: "Category updated successfully", newCategory });
+    } else {
+      res.send("you are not an admin");
+    }
   } catch (err) {
     return next(err);
   }
 };
 
-module.exports = { createCategory, getAllCategories,deleteCategoryById,updateCategoryById };
+//http://localhost:8080/categories/:id
+
+const deleteCategoryById = async (req, res, next) => {
+  try {
+    if (req.authorizedUser.role == "admin") {
+      const categories = await Categories.findById(req.params.id);
+      if (!categories) return next(new AppError("category does not exist"));
+      await Categories.findByIdAndDelete(req.params.id);
+      res.send({ message: "Category deleted successfully" });
+    } else {
+      res.send("you are not an admin");
+    }
+  } catch (err) {
+    return next(err);
+  }
+};
+
+module.exports = {
+  createCategory,
+  getAllCategories,
+  deleteCategoryById,
+  updateCategoryById,
+};
